move methods to models, out of client
