+++ spotify/client.go
@@ -8,7 +8,6 @@ import (
 	"io/ioutil"
 	"net/http"
 	"reflect"
-	"strings"
 	"time"
 
 	redis "github.com/go-redis/redis/v8"
@@ -75,45 +74,6 @@ func checkCache(ctx context.Context, key string) (*[]byte, error) {
 
 	bytes := []byte(val)
 	return &bytes, nil
-}
-
-func getAudioFeatures(ctx context.Context, ids []string) (*AudioFeatures, error) {
-	logger := logging.GetLogger(&ctx)
-	token := ctx.Value(ContextAccessToken)
-	if token == nil {
-		return nil, errors.New("no access token provided")
-	}
-
-	if len(ids) > 100 {
-		logger.WithField("count", len(ids)).Warn("too many tracks passed, reducing to the first 100")
-		ids = ids[:100]
-	}
-
-	url := fmt.Sprint("https://api.spotify.com/v1/audio-features?ids=", strings.Join(ids, ","))
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Add("Authorization", fmt.Sprint("Bearer ", token))
-
-	body, err := makeRequest(ctx, req, true)
-	if err != nil {
-		return nil, err
-	}
-
-	type tempResp struct {
-		Items AudioFeatures `json:"audio_features"`
-	}
-
-	ret := tempResp{}
-	err = json.Unmarshal(*body, &ret)
-	if err != nil {
-		logger.WithField("body", string(*body)).Error(err.Error())
-		return nil, err
-	}
-
-	return &ret.Items, nil
 }
 
 func getChunkOfUserLibraryTracks(ctx context.Context, url string) (Tracks, string, int, error) {
@@ -183,362 +143,6 @@ func getGenres(ctx context.Context) ([]string, error) {
 	}
 
 	return rsp.Genres, nil
-}
-
-func getRecommendations(ctx context.Context, seeds map[string][]string) (*Recommendation, error) {
-	// holy shit, this is actually _really_ configurable.  Come back to this
-	// and explore the possibilities a little more after v1 is out.
-
-	token := ctx.Value(ContextAccessToken)
-	if token == nil {
-		return nil, errors.New("no access token provided")
-	}
-
-	qs := "?"
-	if artists, ok := seeds[KeySeedArtists]; ok {
-		qs = fmt.Sprint(qs, "seed_artists=", strings.Join(artists, ","))
-	}
-
-	if tracks, ok := seeds[KeySeedTracks]; ok {
-		if len(qs) > 1 {
-			qs += "&"
-		}
-
-		qs = fmt.Sprint(qs, "seed_tracks=", strings.Join(tracks, ","))
-	}
-
-	if genres, ok := seeds[KeySeedGenres]; ok {
-		if len(qs) > 1 {
-			qs += "&"
-		}
-
-		qs = fmt.Sprint(qs, "seed_genres=", strings.Join(genres, ","))
-	}
-
-	url := "https://api.spotify.com/v1/recommendations"
-	if len(qs) > 1 {
-		url = fmt.Sprint(url, qs)
-	}
-
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Add("Authorization", fmt.Sprint("Bearer ", token))
-
-	body, err := makeRequest(ctx, req, false)
-	if err != nil {
-		return nil, err
-	}
-
-	type tApiResponse struct {
-		Tracks []struct {
-			Artists []struct {
-				Link string `json:"href"`
-				ID   string `json:"id"`
-				Name string `json:"name"`
-			} `json:"artists"`
-			Duration int64  `json:"duration_ms"`
-			Link     string `json:"href"`
-			ID       string `json:"id"`
-			Name     string `json:"name"`
-		} `json:"tracks"`
-		Seeds []struct {
-			InitialPoolSize    int64  `json:"initialPoolSize"`
-			AfterFilteringSize int64  `json:"afterFilteringSize"`
-			Link               string `json:"href"`
-			ID                 string `json:"id"`
-			Type               string `json:"type"`
-		} `json:"seeds"`
-	}
-
-	var rsp tApiResponse
-	err = json.Unmarshal(*body, &rsp)
-	if err != nil {
-		return nil, err
-	}
-
-	ret := Recommendation{}
-	for _, i := range rsp.Tracks {
-		artists := []string{}
-		for _, j := range i.Artists {
-			artists = append(artists, j.Name)
-		}
-		ret.Tracks = append(ret.Tracks, struct {
-			ID      string   "json:\"id\""
-			Name    string   "json:\"name\""
-			Link    string   "json:\"href\""
-			Artists []string "json:\"artists\""
-		}{
-			ID:      i.ID,
-			Name:    i.Name,
-			Link:    i.Link,
-			Artists: artists,
-		})
-	}
-
-	for _, i := range rsp.Seeds {
-		ret.Seeds = append(ret.Seeds, struct {
-			ID   string "json:\"id\""
-			Link string "json:\"href\""
-			Type string "json:\"type\""
-		}{
-			ID:   i.ID,
-			Link: i.Link,
-			Type: i.Type,
-		})
-	}
-
-	return &ret, nil
-}
-
-func getRelatedArtists(ctx context.Context, id string) (*[]Artist, error) {
-	token := ctx.Value(ContextAccessToken)
-	if token == nil {
-		return nil, errors.New("no access token provided")
-	}
-
-	url := "https://api.spotify.com/v1/artists/%v/related-artists"
-
-	req, err := http.NewRequest("GET", fmt.Sprintf(url, id), nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Add("Authorization", fmt.Sprint("Bearer ", token))
-
-	body, err := makeRequest(ctx, req, false)
-	if err != nil {
-		return nil, err
-	}
-
-	type tRes struct {
-		Artists []struct {
-			Followers struct {
-				Total int64 `json:"total"`
-			} `json:"followers"`
-			Genres     []string `json:"genres"`
-			Link       string   `json:"href"`
-			ID         string   `json:"id"`
-			Name       string   `json:"name"`
-			Popularity int32    `json:"popularity"`
-			Type       string   `json:"type"`
-		} `json:"artists"`
-	}
-
-	rsp := tRes{}
-	err = json.Unmarshal(*body, &rsp)
-	if err != nil {
-		return nil, err
-	}
-
-	ret := []Artist{}
-	for _, i := range rsp.Artists {
-		ret = append(ret, Artist{
-			Genres:     i.Genres,
-			ID:         i.ID,
-			Name:       i.Name,
-			Popularity: i.Popularity,
-		})
-	}
-
-	return &ret, nil
-}
-
-func getTopArtists(ctx context.Context) (*Artists, error) {
-	token := ctx.Value(ContextAccessToken)
-	if token == nil {
-		return nil, errors.New("no access token provided")
-	}
-
-	tr := ctx.Value(ContextTimeRange)
-	strRange := ""
-	if tr != nil {
-		strRange = tr.(string)
-	}
-
-	// TODO: make this limit a param
-	url := "https://api.spotify.com/v1/me/top/artists?limit=25"
-	if len(strRange) > 0 {
-		url += fmt.Sprint("&time_range=", strRange)
-	}
-
-	req, err := http.NewRequest("GET", url, nil)
-
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Add("Authorization", fmt.Sprint("Bearer ", token))
-
-	body, err := makeRequest(ctx, req, true)
-	if err != nil {
-		return nil, err
-	}
-
-	type tempResp struct {
-		Items Artists `json:"items"`
-	}
-
-	var ret tempResp
-	err = json.Unmarshal(*body, &ret)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ret.Items, nil
-}
-
-func getTopTracks(ctx context.Context, limit int32) (Tracks, error) {
-	token := ctx.Value(ContextAccessToken)
-	if token == nil {
-		return nil, errors.New("no access token provided")
-	}
-
-	tr := ctx.Value(ContextTimeRange)
-	strRange := ""
-	if tr != nil {
-		strRange = tr.(string)
-	}
-
-	url := fmt.Sprint("https://api.spotify.com/v1/me/top/tracks?limit=", limit)
-	if len(strRange) > 0 {
-		url += fmt.Sprint("&time_range=", strRange)
-	}
-
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Add("Authorization", fmt.Sprint("Bearer ", token))
-
-	body, err := makeRequest(ctx, req, true)
-	if err != nil {
-		return nil, err
-	}
-
-	type tempResp struct {
-		Items Tracks `json:"items"`
-	}
-
-	var ret tempResp
-	err = json.Unmarshal(*body, &ret)
-	if err != nil {
-		return nil, err
-	}
-
-	return ret.Items, nil
-}
-
-func getTopTracksForArtist(ctx context.Context, id string) (*[]TopTracks, error) {
-	token := ctx.Value(ContextAccessToken)
-	if token == nil {
-		return nil, errors.New("no access token provided")
-	}
-
-	url := fmt.Sprintf("https://api.spotify.com/v1/artists/%v/top-tracks?country=us", id)
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Add("Authorization", fmt.Sprint("Bearer ", token))
-
-	body, err := makeRequest(ctx, req, true)
-	if err != nil {
-		return nil, err
-	}
-
-	type tRes struct {
-		Tracks []struct {
-			ID         string `json:"id"`
-			Name       string `json:"name"`
-			Popularity int64  `json:"popularity"`
-			Artists    []struct {
-				ID   string `json:"id"`
-				Name string `json:"name"`
-				Link string `json:"href"`
-			} `json:"artists"`
-		} `json:"tracks"`
-	}
-
-	rsp := tRes{}
-	err = json.Unmarshal(*body, &rsp)
-	if err != nil {
-		return nil, err
-	}
-
-	ret := []TopTracks{}
-	for _, i := range rsp.Tracks {
-		artists := []string{}
-		for _, j := range i.Artists {
-			artists = append(artists, j.Name)
-		}
-		ret = append(ret, TopTracks{
-			ID:         i.ID,
-			Name:       i.Name,
-			Popularity: i.Popularity,
-			Artists:    artists,
-		})
-	}
-
-	return &ret, nil
-}
-
-func getTracks(ctx context.Context, ids []string) (*Tracks, error) {
-	token := ctx.Value(ContextAccessToken)
-	if token == nil {
-		return nil, errors.New("no access token provided")
-	}
-
-	url := fmt.Sprint("https://api.spotify.com/v1/tracks?ids=", strings.Join(ids, ","))
-
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Add("Authorization", fmt.Sprint("Bearer ", token))
-
-	body, err := makeRequest(ctx, req, true)
-	if err != nil {
-		return nil, err
-	}
-
-	type tempResp struct {
-		Items Tracks `json:"tracks"`
-	}
-
-	var ret tempResp
-	err = json.Unmarshal(*body, &ret)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ret.Items, nil
-}
-
-func getUserLibraryTracks(ctx context.Context) (Tracks, error) {
-	logger := logging.GetLogger(nil)
-	url := "https://api.spotify.com/v1/me/tracks?limit=50&offset=0"
-	more := true
-	ret := []Track{}
-	for more {
-		t, newUrl, tot, err := getChunkOfUserLibraryTracks(ctx, url)
-		if err != nil {
-			logger.Warn(err.Error())
-			more = false
-		}
-		url = newUrl
-
-		ret = append(ret, t...)
-		if tot == len(ret) {
-			more = false
-		}
-	}
-	return ret, nil
 }
 
 func getUserInfo(ctx context.Context) (map[string]string, error) {
